+++ astropy/utils/iers/iers.py
@@ -7,51 +7,11 @@ times.  These are used in `astropy.time` to provide UT1 values.  The polar
 motions are also used for determining earth orientation for
 celestional-to-terrestrial coordinate transformations
 (in `astropy.coordinates`).
-
-By default, IERS B values provided as part of astropy are used, but
-user-downloaded files can be substituted.
-
-Generally, there is no need to invoke the iers classes oneself.  E.g.,
-the default IERS B table is loaded as necessary in `Time`::
-
-    >>> from astropy.time import Time
-    >>> t = Time(['2012-06-30 12:00:00', '2012-06-30 23:59:59',
-    ...           '2012-06-30 23:59:60', '2012-07-01 00:00:00',
-    ...           '2012-07-01 12:00:00'], scale='utc')
-    >>> t.ut1
-    <Time object: scale='ut1' format='iso' value=['2012-06-30 11:59:59.413'
-     '2012-06-30 23:59:58.413' '2012-06-30 23:59:59.413'
-     '2012-07-01 00:00:00.413' '2012-07-01 12:00:00.413']>
-
-But if one is dealing with very recent observations, this does not work::
-
-    >>> t2 = Time.now()
-    >>> t2.ut1  # doctest: +SKIP
-    Traceback (most recent call last):
-    ...
-    IERSRangeError: (some) times are outside of range covered by IERS table.
-
-In this case, one needs to update the IERS B table or use IERS A instead
-(which also has predictions).  In future versions, this may become
-configurable or automatic, but currently it requires handiwork.  For
-`Time`, easiest is to set the `delta_ut1_utc` property directly::
-
-    >>> from astropy.utils.iers import IERS_A
-    >>> iers_a = IERS_A.open('finals2000A.all')    # doctest: +SKIP
-    >>> iers_a.ut1_utc(t2)                         # doctest: +SKIP
-    <Quantity -0.5631187323911956 s>
-    >>> t2.delta_ut1_utc = iers_a.ut1_utc(t2)      # doctest: +SKIP
-    >>> t2.ut1.iso                                 # doctest: +SKIP
-    '2015-03-24 03:33:57.288'
-
-Instead of local copies of IERS files, one can also download them, using
-`iers.IERS_A_URL` and `iers.IERS_B_URL`::
-
-    >>> from astropy.utils.iers import IERS_A, IERS_A_URL
-    >>> iers_a = IERS_A.open(IERS_A_URL)                    # doctest: +SKIP
 """
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
+
+from warnings import warn
 
 try:
     from urlparse import urlparse
@@ -60,24 +20,31 @@ except ImportError:
 
 import numpy as np
 
+from ... import config as _config
+from ... import units as u
 from ...table import Table, QTable
-from ...utils.data import get_pkg_data_filename, download_file
+from ...utils.data import get_pkg_data_filename, download_file, clear_download_cache
+from ...utils.exceptions import AstropyWarning
+from ...tests import disable_internet
 
-__all__ = ['IERS', 'IERS_B', 'IERS_A',
+__all__ = ['Conf', 'conf',
+           'IERS', 'IERS_B', 'IERS_A', 'IERS_Auto',
            'FROM_IERS_B', 'FROM_IERS_A', 'FROM_IERS_A_PREDICTION',
            'TIME_BEFORE_IERS_RANGE', 'TIME_BEYOND_IERS_RANGE',
            'IERS_A_FILE', 'IERS_A_URL', 'IERS_A_README',
            'IERS_B_FILE', 'IERS_B_URL', 'IERS_B_README',
-           'IERSRangeError']
+           'IERSRangeError', 'IERSStaleWarning']
 
 # IERS-A default file name, URL, and ReadMe with content description
 IERS_A_FILE = 'finals2000A.all'
 IERS_A_URL = 'http://maia.usno.navy.mil/ser7/finals2000A.all'
 IERS_A_README = get_pkg_data_filename('data/ReadMe.finals2000A')
+
 # IERS-B default file name, URL, and ReadMe with content description
 IERS_B_FILE = get_pkg_data_filename('data/eopc04_IAU2000.62-now')
 IERS_B_URL = 'http://hpiers.obspm.fr/iers/eop/eopc04/eopc04_IAU2000.62-now'
 IERS_B_README = get_pkg_data_filename('data/ReadMe.eopc04_IAU2000')
+
 # Status/source values returned by IERS.ut1_utc
 FROM_IERS_B = 0
 FROM_IERS_A = 1
@@ -85,8 +52,46 @@ FROM_IERS_A_PREDICTION = 2
 TIME_BEFORE_IERS_RANGE = -1
 TIME_BEYOND_IERS_RANGE = -2
 
-
 MJD_ZERO = 2400000.5
+
+INTERPOLATE_ERROR = """\
+interpolating from IERS_Auto using predictive values that are more
+than {} days old.
+
+Normally you should not see this error because this class
+automatically downloads the latest IERS-A table.  Perhaps you are
+offline?  If you understand what you are doing then this error can be
+suppressed by setting the auto_max_age configuration variable to
+``None``:
+
+  from astropy.utils.iers import conf
+  conf.auto_max_age = None
+"""
+
+class IERSStaleWarning(AstropyWarning):
+    pass
+
+
+class Conf(_config.ConfigNamespace):
+    """
+    Configuration parameters for `astropy.utils.iers`.
+    """
+    auto_download = _config.ConfigItem(
+        True,
+        'Enable auto-downloading of the latest IERS data.  If set to False '
+        'then the local IERS-B file will be used by default. (default=True)')
+    auto_max_age = _config.ConfigItem(
+        30.0,
+        'Maximum age of predictive data before auto-downloading (days, default=30)')
+    iers_auto_url = _config.ConfigItem(
+        IERS_A_URL,
+        'URL for auto-downloading IERS file data')
+
+conf = Conf()
+
+# If internet is off for testing then do not download.  This
+# makes most tests fall back to using the built-in IERS-B table.
+conf.auto_download &= not disable_internet.INTERNET_OFF
 
 
 class IERSRangeError(IndexError):
@@ -111,9 +116,9 @@ class IERS(QTable):
         ----------
         file : str or None
             full local or network path to the ascii file holding IERS data,
-            for passing on to the `read` class methods (further optional
+            for passing on to the ``read`` class methods (further optional
             arguments that are available for some IERS subclasses can be added).
-            If None, use the default location from the `read` class method.
+            If None, use the default location from the ``read`` class method.
         cache : bool
             Whether to use cache. Defaults to False, since IERS files
             are regularly updated.
@@ -124,9 +129,9 @@ class IERS(QTable):
 
         Notes
         -----
-        On the first call in a session, the table will be memoized (in
-        `cls.iers_table`), and further calls to `open` will return this stored
-        table if `file=None` (the default).
+        On the first call in a session, the table will be memoized (in the
+        ``iers_table`` class attribute), and further calls to ``open`` will
+        return this stored table if ``file=None`` (the default).
 
         If a table needs to be re-read from disk, pass on an explicit file
         location or use the (sub-class) close method and re-open.
@@ -135,6 +140,7 @@ class IERS(QTable):
         download_file.
 
         For the IERS class itself, an IERS_B sub-class instance is opened.
+
         """
         if file is not None or cls.iers_table is None:
             if file is not None:
@@ -190,7 +196,7 @@ class IERS(QTable):
             second part of two-part JD (default 0., ignored if jd1 is Time)
         return_status : bool
             Whether to return status values.  If False (default),
-            raise `IERSRangeError` if any time is out of the range covered
+            raise ``IERSRangeError`` if any time is out of the range covered
             by the IERS table.
 
         Returns
@@ -198,12 +204,12 @@ class IERS(QTable):
         ut1_utc : float or float array
             UT1-UTC, interpolated in IERS Table
         status : int or int array
-            Status values (if `return_status`=`True`)::
-            `iers.FROM_IERS_B`
-            `iers.FROM_IERS_A`
-            `iers.FROM_IERS_A_PREDICTION`
-            `iers.TIME_BEFORE_IERS_RANGE`
-            `iers.TIME_BEYOND_IERS_RANGE`
+            Status values (if ``return_status``=``True``)::
+            ``iers.FROM_IERS_B``
+            ``iers.FROM_IERS_A``
+            ``iers.FROM_IERS_A_PREDICTION``
+            ``iers.TIME_BEFORE_IERS_RANGE``
+            ``iers.TIME_BEYOND_IERS_RANGE``
         """
         return self._interpolate(jd1, jd2, ['UT1_UTC'],
                                  self.ut1_utc_source if return_status else None)
@@ -219,7 +225,7 @@ class IERS(QTable):
             second part of two-part JD (default 0., ignored if jd1 is Time)
         return_status : bool
             Whether to return status values.  If False (default),
-            raise `IERSRangeError` if any time is out of the range covered
+            raise ``IERSRangeError`` if any time is out of the range covered
             by the IERS table.
 
         Returns
@@ -229,15 +235,25 @@ class IERS(QTable):
         PM_y : Quantity with angle units
             y component of polar motion for the requested times
         status : int or int array
-            Status values (if `return_status`=`True`)::
-            `iers.FROM_IERS_B`
-            `iers.FROM_IERS_A`
-            `iers.FROM_IERS_A_PREDICTION`
-            `iers.TIME_BEFORE_IERS_RANGE`
-            `iers.TIME_BEYOND_IERS_RANGE`
+            Status values (if ``return_status``=``True``)::
+            ``iers.FROM_IERS_B``
+            ``iers.FROM_IERS_A``
+            ``iers.FROM_IERS_A_PREDICTION``
+            ``iers.TIME_BEFORE_IERS_RANGE``
+            ``iers.TIME_BEYOND_IERS_RANGE``
         """
         return self._interpolate(jd1, jd2, ['PM_x', 'PM_y'],
                                  self.pm_source if return_status else None)
+
+    def _check_interpolate_indices(self, indices_orig, indices_clipped, max_input_mjd):
+        """
+        Check that the indices from interpolation match those after clipping
+        to the valid table range.  This method gets overridden in the IERS_Auto
+        class because it has different requirements.
+        """
+        if np.any(indices_orig != indices_clipped):
+            raise IERSRangeError('(some) times are outside of range covered '
+                                 'by IERS table.')
 
     def _interpolate(self, jd1, jd2, columns, source=None):
         mjd, utc = self.mjd_utc(jd1, jd2)
@@ -246,14 +262,18 @@ class IERS(QTable):
         if is_scalar:
             mjd = np.array([mjd])
             utc = np.array([utc])
+
+        self._refresh_table_as_needed(mjd)
+
         # For typical format, will always find a match (since MJD are integer)
         # hence, important to define which side we will be; this ensures
         # self['MJD'][i-1]<=mjd<self['MJD'][i]
         i = np.searchsorted(self['MJD'].value, mjd, side='right')
+
         # Get index to MJD at or just below given mjd, clipping to ensure we
         # stay in range of table (status will be set below for those outside)
-        i1 = np.clip(i, 1, len(self)-1)
-        i0 = i1-1
+        i1 = np.clip(i, 1, len(self) - 1)
+        i0 = i1 - 1
         mjd_0, mjd_1 = self['MJD'][i0].value, self['MJD'][i1].value
         results = []
         for column in columns:
@@ -268,6 +288,10 @@ class IERS(QTable):
             # interpolation and correction for tidal effects;
             # http://maia.usno.navy.mil/iers-gaz13)
             val = val_0 + (mjd - mjd_0 + utc) / (mjd_1 - mjd_0) * d_val
+
+            # Do not extrapolate outside range, instead just propagate last values.
+            val[i == 0] = self[column][0]
+            val[i == len(self)] = self[column][-1]
 
             if is_scalar:
                 val = val[0]
@@ -285,11 +309,16 @@ class IERS(QTable):
             results.append(status)
             return results
         else:
-            # Not returning status, so raise an exception for out-of-range
-            if np.any(i1 != i):
-                raise IERSRangeError('(some) times are outside of range covered '
-                                 'by IERS table.')
+            self._check_interpolate_indices(i1, i, np.max(mjd))
             return results[0] if len(results) == 1 else results
+
+    def _refresh_table_as_needed(self, mjd):
+        """
+        Potentially update the IERS table in place depending on the requested
+        time values in ``mdj`` and the time span of the table.  The base behavior
+        is not to update the table.  ``IERS_Auto`` overrides this method.
+        """
+        pass
 
     def ut1_utc_source(self, i):
         """Source for UT1-UTC.  To be overridden by subclass."""
@@ -298,6 +327,18 @@ class IERS(QTable):
     def pm_source(self, i):
         """Source for polar motion.  To be overridden by subclass."""
         return np.zeros_like(i)
+
+    @property
+    def time_now(self):
+        """
+        Property to provide the current time, but also allow for explicitly setting
+        the _time_now attribute for testing purposes.
+        """
+        from astropy.time import Time
+        try:
+            return self._time_now
+        except:
+            return Time.now()
 
 
 class IERS_A(IERS):
@@ -309,18 +350,42 @@ class IERS_A(IERS):
     Notes
     -----
     The IERS A file is not part of astropy.  It can be downloaded from
-    `iers.IERS_A_URL`.  See `iers.__doc__` for instructions on how to use
-    it in `Time`, etc.
+    ``iers.IERS_A_URL``.  See ``iers.__doc__`` for instructions on how to use
+    it in ``Time``, etc.
     """
 
     iers_table = None
 
-    def __init__(self, table):
-        """Initialize an IERS-A table that is already read in.
-        Use read or open class methods to read it from disk.
-
-        Combines UT1-UTC values, taking UT1_UTC_B if available, else UT1_UTC_A
+    @staticmethod
+    def _combine_a_b_columns(iers_a):
         """
+        Return a new table with appropriate combination of IERS_A and B columns.
+        """
+        # IERS A has some rows at the end that hold nothing but dates & MJD
+        # presumably to be filled later.  Exclude those a priori -- there
+        # should at least be a predicted UT1-UTC and PM!
+        table = iers_a[~iers_a['UT1_UTC_A'].mask &
+                       ~iers_a['PolPMFlag_A'].mask]
+
+        # IERS-A has IERS-B values included, but for reasons unknown these do not
+        # match the latest IERS-B values.  See comments in #4436 for details.  So
+        # here we use the bundled astropy IERS-B table to overwrite the values
+        # in the downloaded IERS-A table.
+        iers_b = IERS_B.open()
+        # IERS-B starts before IERS-A (the finals2000A.all version) so fix that.
+        i0 = np.searchsorted(iers_b['MJD'].value, table['MJD'][0])
+        iers_b = iers_b[i0:]
+        n_iers_b = len(iers_b)
+        # If there is overlap then replace IERS-A values from available IERS-B
+        if n_iers_b > 0:
+            # Sanity check that we are overwriting the correct values
+            if not np.allclose(table['MJD'][:n_iers_b], iers_b['MJD'].value):
+                raise ValueError('unexpected mismatch when copying IERS-B values into IERS-A table')
+            # Finally do the overwrite
+            table['UT1_UTC_B'][:n_iers_b] = iers_b['UT1_UTC'].value
+            table['PM_X_B'][:n_iers_b] = iers_b['PM_x'].value
+            table['PM_Y_B'][:n_iers_b] = iers_b['PM_y'].value
+
         # Run np.where on the data from the table columns, since in numpy 1.9
         # it otherwise returns an only partially initialized column.
         table['UT1_UTC'] = np.where(table['UT1_UTC_B'].mask,
@@ -343,36 +408,51 @@ class IERS_A(IERS):
         table['PolPMFlag'] = np.where(table['PM_X_B'].mask,
                                       table['PolPMFlag_A'].data,
                                       'B')
-        # Fill any masked values, and convert to a QTable.
-        super(IERS_A, self).__init__(table.filled())
+
+        # Get the table index for the first row that has predictive values
+        # PolPMFlag_A  IERS (I) or Prediction (P) flag for
+        #              Bull. A polar motion values
+        # UT1Flag_A    IERS (I) or Prediction (P) flag for
+        #              Bull. A UT1-UTC values
+        is_predictive = (table['UT1Flag_A'] == 'P') | (table['PolPMFlag_A'] == 'P')
+        table.meta['predictive_index'] = np.min(np.flatnonzero(is_predictive))
+        table.meta['predictive_mjd'] = table['MJD'][table.meta['predictive_index']]
+
+        return table
 
     @classmethod
-    def read(cls, file=IERS_A_FILE, readme=None):
+    def read(cls, file=None, readme=None):
         """Read IERS-A table from a finals2000a.* file provided by USNO.
 
         Parameters
         ----------
         file : str
             full path to ascii file holding IERS-A data
-            (default: `iers.IERS_A_FILE`)
+            (default: ``iers.IERS_A_FILE``)
         readme : str
             full path to ascii file holding CDS-style readme
-            (default: package version, `iers.IERS_A_README`)
+            (default: package version, ``iers.IERS_A_README``)
 
         Returns
         -------
-        `IERS_A` class instance
+        ``IERS_A`` class instance
         """
+        if file is None:
+            file = IERS_A_FILE
         if readme is None:
             readme = IERS_A_README
+
         # Read in as a regular Table, including possible masked columns.
         # Columns will be filled and converted to Quantity in cls.__init__.
         iers_a = Table.read(file, format='cds', readme=readme)
-        # IERS A has some rows at the end that hold nothing but dates & MJD
-        # presumably to be filled later.  Exclude those a priori -- there
-        # should at least be a predicted UT1-UTC and PM!
-        return cls(iers_a[~iers_a['UT1_UTC_A'].mask &
-                          ~iers_a['PolPMFlag_A'].mask])
+
+        # Combine the A and B data for UT1-UTC and PM columns
+        table = cls._combine_a_b_columns(iers_a)
+        table.meta['data_path'] = file
+        table.meta['readme_path'] = readme
+
+        # Fill any masked values, and convert to a QTable.
+        return cls(table.filled())
 
     def ut1_utc_source(self, i):
         """Set UT1-UTC source flag for entries in IERS table"""
@@ -398,9 +478,8 @@ class IERS_B(IERS):
 
     Notes
     -----
-    If the package IERS B file (`iers.IERS_B_FILE`) is out of date, a new
-    version can be downloaded from `iers.IERS_B_URL`.  See `iers.__doc__`
-    for instructions on how to use it in `Time`, etc.
+    If the package IERS B file (```iers.IERS_B_FILE``) is out of date, a new
+    version can be downloaded from ``iers.IERS_B_URL``.
     """
 
     iers_table = None
@@ -413,16 +492,16 @@ class IERS_B(IERS):
         ----------
         file : str
             full path to ascii file holding IERS-B data
-            (default: package version, `iers.IERS_B_FILE`)
+            (default: package version, ``iers.IERS_B_FILE``)
         readme : str
             full path to ascii file holding CDS-style readme
-            (default: package version, `iers.IERS_B_README`)
+            (default: package version, ``iers.IERS_B_README``)
         data_start : int
             starting row (default: 14, appropriate for standard IERS files)
 
         Returns
         -------
-        `IERS_B` class instance
+        ``IERS_B`` class instance
         """
         if file is None:
             file = IERS_B_FILE
@@ -442,6 +521,150 @@ class IERS_B(IERS):
     def pm_source(self, i):
         """Set PM source flag for entries in IERS table"""
         return np.ones_like(i) * FROM_IERS_B
+
+class IERS_Auto(IERS_A):
+    """
+    Provide most-recent IERS data and automatically handle downloading
+    of updated values as necessary.
+    """
+    iers_table = None
+
+    @classmethod
+    def open(cls):
+        """If the configuration setting ``astropy.utils.iers.conf.auto_download``
+        is set to True (default), then open a recent version of the IERS-A
+        table with predictions for UT1-UTC and polar motion out to
+        approximately one year from now.  If the available version of this file
+        is older than ``astropy.utils.iers.conf.auto_max_age`` days old
+        (or non-existent) then it will be downloaded over the network and cached.
+
+        If the configuration setting ``astropy.utils.iers.conf.auto_download``
+        is set to False then ``astropy.utils.iers.IERS()`` is returned.  This
+        is normally the IERS-B table that is supplied with astropy.
+
+        On the first call in a session, the table will be memoized (in the
+        ``iers_table`` class attribute), and further calls to ``open`` will
+        return this stored table.
+
+        Returns
+        -------
+        `~astropy.table.QTable` instance with IERS (Earth rotation) data columns
+
+        """
+        if not conf.auto_download:
+            cls.iers_table = IERS.open()
+
+        if cls.iers_table is not None:
+            return cls.iers_table
+
+        try:
+            filename = download_file(conf.iers_auto_url, cache=True)
+        except Exception as err:
+            # Issue a warning here, perhaps user is offline.  An exception
+            # will be raised downstream when actually trying to interpolate
+            # predictive values.
+            warn(AstropyWarning('failed to download {}, using local IERS-B: {}'
+                                .format(conf.iers_auto_url, str(err))))
+            cls.iers_table = IERS.open()
+            return cls.iers_table
+
+        cls.iers_table = cls.read(file=filename)
+        cls.iers_table.meta['data_url'] = str(conf.iers_auto_url)
+
+        return cls.iers_table
+
+    def _check_interpolate_indices(self, indices_orig, indices_clipped, max_input_mjd):
+        """Check that the indices from interpolation match those after clipping to the
+        valid table range.  The IERS_Auto class is exempted as long as it has
+        sufficiently recent available data so the clipped interpolation is
+        always within the confidence bounds of current Earth rotation
+        knowledge.
+        """
+        predictive_mjd = self.meta['predictive_mjd']
+
+        # See explanation in _refresh_table_as_needed for these conditions
+        auto_max_age = (conf.auto_max_age if conf.auto_max_age is not None
+                        else np.finfo(np.float).max)
+        if (max_input_mjd > predictive_mjd and
+                self.time_now.mjd - predictive_mjd > auto_max_age):
+            raise ValueError(INTERPOLATE_ERROR)
+
+    def _refresh_table_as_needed(self, mjd):
+        """Potentially update the IERS table in place depending on the requested
+        time values in ``mjd`` and the time span of the table.
+
+        For IERS_Auto the behavior is that the table is refreshed from the IERS
+        server if both the following apply:
+
+        - Any of the requested IERS values are predictive.  The IERS-A table
+          contains predictive data out for a year after the available
+          definitive values.
+        - The first predictive values are at least ``conf.auto_max_age days`` old.
+          In other words the IERS-A table was created by IERS long enough
+          ago that it can be considered stale for predictions.
+        """
+        max_input_mjd = np.max(mjd)
+        now_mjd = self.time_now.mjd
+
+        # IERS-A table contains predictive data out for a year after
+        # the available definitive values.
+        fpi = self.meta['predictive_index']
+        predictive_mjd = self.meta['predictive_mjd']
+
+
+        # Update table in place if necessary
+        auto_max_age = (conf.auto_max_age if conf.auto_max_age is not None
+                        else np.finfo(np.float).max)
+
+        # If auto_max_age is smaller than IERS update time then repeated downloads may
+        # occur without getting updated values (giving a IERSStaleWarning).
+        if auto_max_age < 10:
+            raise ValueError('IERS auto_max_age configuration value must be larger than 10 days')
+
+        if (max_input_mjd > predictive_mjd and
+               now_mjd - predictive_mjd > auto_max_age):
+
+            # Get the latest version
+            try:
+                clear_download_cache(conf.iers_auto_url)
+                filename = download_file(conf.iers_auto_url, cache=True)
+            except Exception as err:
+                # Issue a warning here, perhaps user is offline.  An exception
+                # will be raised downstream when actually trying to interpolate
+                # predictive values.
+                warn(AstropyWarning('failed to download {}: {}.\nA coordinate or time-related '
+                                    'calculation might be compromised or fail because the dates are '
+                                    'not covered by the available IERS file.  See the '
+                                    '"IERS data access" section of the astropy documentation '
+                                    'for additional information on working offline.'
+                                    .format(conf.iers_auto_url, str(err))))
+                return
+
+            new_table = self.__class__.read(file=filename)
+
+            # New table has new values?
+            if new_table['MJD'][-1] > self['MJD'][-1]:
+                # Replace *replace* current values from the first predictive index through
+                # the end of the current table.  This replacement is much faster than just
+                # deleting all rows and then using add_row for the whole duration.
+                new_fpi = np.searchsorted(new_table['MJD'].value, predictive_mjd, side='right')
+                n_replace = len(self) - fpi
+                self[fpi:] = new_table[new_fpi:new_fpi + n_replace]
+
+                # Sanity check for continuity
+                if new_table['MJD'][new_fpi + n_replace] - self['MJD'][-1] != 1.0 * u.d:
+                    raise ValueError('unexpected gap in MJD when refreshing IERS table')
+
+                # Now add new rows in place
+                for row in new_table[new_fpi + n_replace:]:
+                    self.add_row(row)
+
+                self.meta.update(new_table.meta)
+            else:
+                warn(IERSStaleWarning(
+                    'IERS_Auto predictive values are older than {} days but downloading '
+                    'the latest table did not find newer values'.format(conf.auto_max_age)))
+
 
 # by default for IERS class, read IERS-B table
 IERS.read = IERS_B.read